refactor: move outgoing models to domain package
