+++ src/main/kotlin/se/uulm/snowballr/backend/model/outgoing/project/ProjectDecisionStatistics.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.model.outgoing
+package se.uulm.snowballr.backend.model.outgoing.project
 
 import snowballr.ProjectOuterClass
