+++ src/main/kotlin/se/uulm/snowballr/backend/model/outgoing/project/ProjectInformation.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.model.outgoing
+package se.uulm.snowballr.backend.model.outgoing.project
 
 import com.google.protobuf.timestamp
 import snowballr.ProjectOuterClass
